Track repeated requests using iterator position instead of index table
The index table does not work if using custom matchers because it relies
on the request hash, which completely ignores matchers. Instead we can
let the storage iterator keep it's position and just return the next
recording if it matches. If it doesn't match we don't advance the iterator.
In addition to matching correctly this is more performant because we don't
keep checking recordings we've already played to see if they match.